inventory_server/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
gunicorn==23.0.0

